refactor: split util
