+++ src/comments/HTML5Comment.ts
@@ -9,7 +9,6 @@ import type {
   measureTextInput,
   measureTextResult,
 } from "@/@types/";
-import { BaseComment } from "@/comments/BaseComment";
 import { imageCache } from "@/contexts/";
 import { config, options } from "@/definition/config";
 import { CanvasRenderingContext2DError } from "@/errors/";
@@ -19,8 +18,15 @@ import {
   getLineHeight,
   measure,
 } from "@/nico";
-import { getConfig, getStrokeColor, parseFont } from "@/util";
-import { isLineBreakResize, parseCommandAndNicoScript } from "@/utils/comment";
+import {
+  getConfig,
+  getStrokeColor,
+  isLineBreakResize,
+  parseCommandAndNicoScript,
+  parseFont,
+} from "@/utils";
+
+import { BaseComment } from "./BaseComment";
 
 class HTML5Comment extends BaseComment {
   override readonly pluginName: string = "HTML5Comment";
