+++ src/utils/index.ts
@@ -1,0 +1,7 @@
+export * from "./array";
+export * from "./color";
+export * from "./comment";
+export * from "./commentArt";
+export * from "./config";
+export * from "./flash";
+export * from "./sort";
